+++ app/css/application.postcss.css
@@ -37,6 +37,7 @@
 @import "./components/site-update";
 @import "./components/site-footer";
 @import "./components/site-header";
+@import "./components/nav-bar";
 @import "./components/community-rank-tag";
 @import "./components/track-breadcrumbs";
 @import "./components/contributing/header";
