+++ app/css/components/site-header.css
@@ -255,112 +255,6 @@ body.namespace-.controller-insiders {
                 width: 16px;
             }
         }
-
-        .nav-element {
-            @apply relative;
-
-            .arrow {
-                --a-size: 15px;
-                @apply absolute hidden -z-1;
-                width: var(--a-size);
-                height: var(--a-size);
-                bottom: 0;
-                left: 50%;
-                background-color: var(--backgroundColorB);
-                clip-path: polygon(0% 0%, 100% 0%, 0% 100%);
-                transform: translateX(-50%) rotate(45deg);
-                border-radius: 3px;
-                vertical-align: middle;
-            }
-
-            span {
-                @apply p-24;
-                &:hover + .nav-element-dropdown,
-                &:focus-within + .nav-element-dropdown {
-                    ul li {
-                        &:first-child {
-                            .nav-dropdown-element {
-                                @apply bg-backgroundColorA;
-                            }
-                            .nav-dropdown-view {
-                                display: inherit;
-                            }
-                        }
-                    }
-                }
-            }
-            &:hover,
-            &:focus-within {
-                .nav-element-dropdown {
-                    display: flex;
-                }
-                .arrow {
-                    display: inherit;
-                }
-            }
-            .nav-element-dropdown {
-                --nd-height: 380px;
-                --nd-width: 700px;
-                --nd-backgroundColor: var(--backgroundColorB);
-                --nd-highlightColor: var(--backgroundColorA);
-                --dropdown-offset: 0px;
-
-                height: var(--nd-height);
-                width: var(--nd-width);
-                @apply absolute hidden top-64 z-menu;
-                left: var(--dropdown-offset);
-                outline: 8px solid var(--nd-backgroundColor);
-                @apply bg-backgroundColorB shadow-base rounded-8;
-
-                ul {
-                    @apply flex flex-col items-baseline;
-                    @apply mr-0;
-                    @apply relative;
-                    width: calc(var(--nd-width) * calc(2 / 5));
-
-                    li {
-                        @apply pr-8 w-100;
-                        .nav-dropdown-view {
-                            @apply absolute hidden p-32 rounded-8;
-                            background: var(--nd-highlightColor);
-                            height: var(--nd-height);
-                            top: var(--nd-height);
-                            width: calc(var(--nd-width) * calc(3 / 5));
-                            transform: translate(100%, -100%);
-                            right: 0;
-                        }
-
-                        &:hover,
-                        &:focus-within {
-                            .nav-dropdown-view {
-                                display: inherit;
-                            }
-
-                            .nav-dropdown-element {
-                                background: var(--nd-highlightColor);
-                            }
-                        }
-                        .nav-dropdown-element {
-                            @apply p-16 rounded-8 flex gap-8 items-center;
-
-                            .c-icon {
-                                @apply filter-textColor6;
-                                height: 24px;
-                                width: 24px;
-                            }
-
-                            h6 {
-                                @apply text-14 text-textColor1 leading-160;
-                            }
-
-                            p {
-                                @apply text-p-base text-13;
-                            }
-                        }
-                    }
-                }
-            }
-        }
     }
 
     & nav.signed-in {
