+++ app/helpers/view_components/nav_helpers/all.rb
@@ -1,0 +1,9 @@
+module ViewComponents
+  module NavHelpers
+    module All
+      include GenericNavHelper
+      include NavDropdownHelper
+      include NavSubmenus
+    end
+  end
+end
