+++ app/javascript/packs/application.tsx
@@ -470,10 +470,13 @@ import type { AutomationLockedTooltipProps } from '../components/tooltips/Automa
 import type { DigDeeperProps } from '@/components/track/DigDeeper'
 import type { ChartData } from '@/components/impact/Chart'
 import { InsidersStatusData } from '../components/insiders/InsiderStatus.js'
+import { handleNavbarFocus } from '@/utils/navbar-focus-handler.js'
 
 document.addEventListener('turbo:load', () => {
   highlightAll()
 })
+
+handleNavbarFocus()
 
 // object.entries polyfill
 if (!Object.entries) {
